Chat overlay auto login. Make global ChatOverlayOption.
