+++ libs/overlay/src/index.ts
@@ -1,4 +1,4 @@
 export * from './lib/ChatOverlay';
 export * from './lib/ChatOverlayManager';
 
-export { Feature } from '@epam/ai-dial-shared';
+export { Feature, type ChatOverlayOptions } from '@epam/ai-dial-shared';
